PdfWeb/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pypdf==4.3.1
pdf2image==1.17.0
Pillow==10.4.0
python-docx==1.1.2
pdfminer.six==20240706
reportlab>=4.0.0
python-multipart==0.0.9
